requirements.txt
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
decorator==4.4.2
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
idna==2.10
imageio-ffmpeg==0.4.4
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
#Jinja2==2.11.3
joblib==1.0.1
Js2Py==0.71
jsonify==0.5
MarkupSafe==1.1.1
moviepy==1.0.3
packaging==20.9
parso==0.8.2
pdf2image==1.16.0
Pillow==2.2.2
proglog==0.1.9
prompt-toolkit==3.0.18
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytesseract==0.3.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
SpeechRecognition==3.8.1
tesseract==0.1.3
threadpoolctl==2.2.0
tokenizers==0.10.2
tornado==6.1
tqdm==4.60.0
transformers==4.5.1
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.1
zope.interface==5.4.0
PyPDF2